feat: Add language-aware navigation to film page
This commit adds language-aware navigation to the film page in the `src/routes/admin/film/[id]/+page.svelte` file. It imports the `languageAwareGoto` function from the `languageAware` utility module and uses it to navigate to the create show page for the selected film. The previous version of toggling the show form is commented out.